EWPP-856: Add block test.
